feat(healing): per-round numbers in chart hover card + enemy effects as right-hand panel (match DPS pattern)
