+++ src/constants/changelog.ts
@@ -11,7 +11,7 @@ export const UNRELEASED_CHANGES: string[] = [
     "Major DPS Calculator engine upgrade: ship skills are auto-parsed into an editable ability model, and the simulation now runs a real combat engine — speed-ordered turns for team ships and the enemy (with a per-config Turn Order strip), live conditional gates, application-time debuff landing with persistent windows, buff-family tier rules, charge cadences with passive auras, and reactive triggers (charge-on-inflict, crit-triggered effects, start-of-round buffs) firing from real combat events. Includes numerous accuracy fixes (Valerian, Hemlock, Oleander, Enforcer, and persistent stacking debuffs like Defense Shred that climb per hit and never expire). Team ships are now full combat actors too: pick a ship and its real skills are parsed and simulated — own stats and affinity, real damage shown separately as Team damage (reduces enemy HP without polluting the attacker's DPS comparison), DoTs and debuffs on their actual turns, reactive triggers, and ally-wide buffs and charge grants for the whole team. With a team configured, the round chart adds a dashed combined-total line carrying the kill mark at the round the enemy actually dies, with a dashed legend swatch to tell team lines apart from the solid attacker lines. Combat Settings is now split into separate Enemy Settings and Team sections you can open independently. Multi-hit skills (such as Enforcer's three-hit active) now crit-check each hit individually — crit-triggered effects like Defense Shred fire once per critting hit rather than once per turn, matching in-game behaviour. Ships with extra-action passives (Liberator, Nuqtu, Sustainer, Tormenter, Tygr) take a full additional turn each round, re-entering the turn order at their Speed position; rounds where an extra turn occurs show a '+N extra turn' line in the round chart tooltip. Ships that react when an ally lands a DoT with a critical hit (Crocus, which inflicts Corrosion II on that enemy) now trigger correctly in team simulations.",
     "Rebuilt Healing Calculator on the same deterministic combat engine as the DPS Calculator. Heal, shield, and cleanse abilities are now auto-parsed from each healer's skill text into an editable ability model — open the Skill Editor to tune heal percentages, scaling stat, shield amounts, and cleanse counts per skill. Pick a heal target (the healer itself, another ship, or manual stats) and add up to four enemy attackers — real ships whose charged skills produce realistic damage spikes, or quick manual stat lines — that bombard the target round by round. Results separate effective healing from overheal, track a shield absorption pool that drains before HP (capped at the target's max HP), tick heal-over-time effects (Repair Over Time) at the holder's turn, and resolve crit heals on their own crit schedule. A survival metric reports how many rounds the target stays up — or the round it is destroyed. Team healers contribute their own heals, shields, and cleanses, and reactive heal triggers (such as Pallas reacting to critically repairing an ally) fire from real combat events. Repairs without an explicit recipient on pure support skills (Hermes, Mender, Salvation, Makoli and 13 more) correctly heal the targeted ally rather than the caster — while damage-skill repair riders and self-conditional heals like Meatshield's stay self-heals. Passive repairs triggered by an ally taking damage, or by a Supporter cleansing a debuff (Cultivator), likewise heal that ally, while a Defender's cleanse-repair (Morao) stays a self-heal. Ships that repair or gain shields based on a percentage of damage dealt or taken — lifesteal-style sustain (including from damage-over-time ticks), burst-triggered heals, and shield-on-hit — are now simulated too. Ships that gain a shield when an enemy is debuffed (APEX) or when they apply Stasis (Defiant) now react to their own actions, and ships that grant themselves a repair buff alongside another effect (Hermes' Everliving Regeneration) now have that buff picked up automatically. Team healers expose an editable Heal Modifier so their heal output reflects their real heal-boost stat. Skill auto-parsing now reads only a ship's active refit passive, so it no longer double-counts buffs from inactive refit tiers. Enemy attackers now act as full combat actors on the same pipeline as your ships — their affinity is matched against the heal target's affinity to scale incoming damage, they apply their real debuffs and DoTs to the tank (which tick for damage each round), and buff themselves with their own abilities. A new Enemy Effects panel shows, round by round, the effects grouped by each enemy attacker (labelled with its ship name): the self-buffs active on that enemy and the debuffs or DoTs it has landed on the target. Like the DPS engine, everything is fully deterministic — identical inputs always produce identical results.",
     'Cleaned up the Damage Deconstruction calculator with a side-by-side attacker/defender layout and clearer result cards.',
-    'Healing Calculator: added a per-round Healer Output overview (mirrors the DPS round breakdown). For the best config it lists, round by round, the direct heal, heal-over-time, shield granted, effective vs overheal, cleanses, and the incoming damage / target HP% it faced — with charged and crit badges.',
+    "Healing Calculator: the per-round healer output now lives in the timeline chart's hover card (matching the DPS calculator) — hover any round to see that round's direct heal, heal-over-time, shield granted, effective vs overheal, cleanses, and the incoming damage / target HP% it faced, with charged and crit badges. The Enemy Effects breakdown is now a hover-gated panel beside the chart that shows the hovered round's per-enemy self-buffs and the debuffs or DoTs landed on the target.",
 ];
 
 export const CHANGELOG: ChangelogEntry[] = [
