+++ src/pages/DocumentationPage.tsx
@@ -2409,10 +2409,11 @@ const DocumentationPage: React.FC = () => {
                                         damage each round — and buff themselves with their own
                                         abilities) or quick manual attack/defense lines. Each enemy
                                         attacker has an affinity selector matched against the
-                                        target&apos;s affinity to scale their incoming damage. An
-                                        Enemy Effects panel shows, round by round, which self-buffs
-                                        are active on each enemy and which debuffs or DoTs they have
-                                        landed on the target.
+                                        target&apos;s affinity to scale their incoming damage. A
+                                        hover-gated Enemy Effects panel beside the timeline chart
+                                        shows, for the hovered round, which self-buffs are active on
+                                        each enemy and which debuffs or DoTs they have landed on the
+                                        target.
                                     </p>
                                     <p className="text-theme-text mb-2">
                                         <span className="text-primary">
@@ -2468,10 +2469,11 @@ const DocumentationPage: React.FC = () => {
                                         <span className="text-primary">
                                             Per-Round Healer Overview:
                                         </span>{' '}
-                                        A round-by-round breakdown of the best config&apos;s output
-                                        — direct heal, heal-over-time, shield granted, effective vs
+                                        Hover any round on the timeline chart to see that
+                                        round&apos;s healer output in the chart&apos;s hover card —
+                                        direct heal, heal-over-time, shield granted, effective vs
                                         overheal, cleanses, and the incoming damage / target HP% it
-                                        faced each round, with charged and crit badges.
+                                        faced — with charged and crit badges.
                                     </p>
                                 </div>
